+++ src/api/projects/protected_branches.rs
@@ -9,6 +9,8 @@
 //! These endpoints are used for querying a project's protected branches.
 
 mod protect;
+mod protected_branch;
+mod protected_branches;
 mod unprotect;
 
 pub use self::protect::ProtectBranch;
@@ -18,3 +20,9 @@ pub use self::protect::ProtectedAccessLevel;
 
 pub use self::unprotect::UnprotectBranch;
 pub use self::unprotect::UnprotectBranchBuilder;
+
+pub use self::protected_branch::ProtectedBranch;
+pub use self::protected_branch::ProtectedBranchBuilder;
+
+pub use self::protected_branches::ProtectedBranches;
+pub use self::protected_branches::ProtectedBranchesBuilder;